Asociar Convocatorias a Proyectos

+++ src/main/java/edu/unicauca/SivriBackendApp/core/proyecto/infraestructure/persistence/jpaRepository/IProyectoRepository.java
@@ -1,9 +1,10 @@
 package edu.unicauca.SivriBackendApp.core.proyecto.infraestructure.persistence.jpaRepository;
 
+
 import edu.unicauca.SivriBackendApp.core.proyecto.infraestructure.persistence.jpaEntity.ProyectoEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IProyectoRepository extends JpaRepository<ProyectoEntity, Integer>{
+public interface IProyectoRepository extends JpaRepository<ProyectoEntity, Long> {
 }
